requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
torch==2.7.1
torchvision==0.22.1
pillow==11.3.0
pymongo==4.13.2
python-multipart==0.0.20
ultralytics==8.3.1
pandas==2.3.1
python-dotenv==1.0.1
google-cloud-storage==2.19.0
onnx==1.17.0
onnxruntime==1.22.1
numpy==2.2.6
opencv-python==4.12.0.88